+++ R/helpers-tests.R
@@ -38,8 +38,85 @@ get_testdata <- function() {
   )
 }
 
+#' get_testdata_combo
+#'
+#' Function to obtain data from gDRtestData and prepare for unit tests
+#'
+#' @examples 
+#' get_testdata_combo()
+#'
+#' @keywords test_helpers
+#' @return list with drugs, cell_lines, raw_data and assay_data
+#'
+#' @export
+get_testdata_combo <- function() {
+  
+  mae <- gDRutils::get_synthetic_data("finalMAE_combo_matrix")
+  raw_data <- gDRutils::convert_mae_assay_to_dt(mae, "Metrics")
+  drug_names <- unique(raw_data$DrugName)
+  cell_line_names <- unique(raw_data$CellLineName)
+  
+  # getting first occurrence of drug_names for each cell_line to avoid aggregation
+  dt <- raw_data[, .SD[1], by = c("DrugName", "CellLineName")]
+  data.table::setnames(dt,
+                       c("CellLineName", "DrugName", "drug_moa", "DrugName_2", "drug_moa_2",
+                         "x_inf", "x_0", "xc50", "h", "r2", "x_sd_avg", "x_mean", "x_AOC_range", 
+                         "x_max", "maxlog10Concentration"),
+                       c("Cell Line Name", "Drug Name", "Drug MOA", "Drug Name 2", "Drug MOA2",
+                         "GR Inf", "GR 0", "GEC50", "h GR", "E Inf", "E0", "EC50", "h RV", 
+                         "GR Max", "Concentration"))
+  dt$GR50 <- dt$EC50
+  dt$IC50 <- dt$EC50
+  dt$`E Max` <- dt$EC50
+  dt$`GR value` <- dt$EC50
+  
+  list(
+    drug_names = drug_names,
+    cell_line_names = cell_line_names,
+    dt = dt,
+    raw_data = raw_data
+  )
+}
 
-
+#' get_testdata_codilution
+#'
+#' Function to obtain data from gDRtestData and prepare for unit tests
+#'
+#' @examples 
+#' get_testdata_codilution()
+#'
+#' @keywords test_helpers
+#' @return list with drugs, cell_lines, raw_data and assay_data
+#'
+#' @export
+get_testdata_codilution <- function() {
+  
+  mae <- gDRutils::get_synthetic_data("finalMAE_combo_codilution_small")
+  raw_data <- gDRutils::convert_mae_assay_to_dt(mae, "Metrics")
+  drug_names <- unique(raw_data$DrugName)
+  cell_line_names <- unique(raw_data$CellLineName)
+  
+  # getting first occurrence of drug_names for each cell_line to avoid aggregation
+  dt <- raw_data[, .SD[1], by = c("DrugName", "CellLineName")]
+  data.table::setnames(dt,
+                       c("CellLineName", "DrugName", "drug_moa", "DrugName_2", "drug_moa_2",
+                         "x_inf", "x_0", "xc50", "h", "r2", "x_sd_avg", "x_mean", "x_AOC_range", 
+                         "x_max", "maxlog10Concentration", "Concentration_2"),
+                       c("Cell Line Name", "Drug Name", "Drug MOA", "Drug Name 2", "Drug MOA2",
+                         "GR Inf", "GR 0", "GEC50", "h GR", "E Inf", "E0", "EC50", "h RV", 
+                         "GR Max", "Concentration", "Concentration 2"))
+  dt$GR50 <- dt$EC50
+  dt$IC50 <- dt$EC50
+  dt$`E Max` <- dt$EC50
+  dt$`GR value` <- dt$EC50
+  
+  list(
+    drug_names = drug_names,
+    cell_line_names = cell_line_names,
+    dt = dt,
+    raw_data = raw_data
+  )
+}
 
 #' gen_synthetic_data
 #'