+++ NEWS.md
@@ -1,5 +1,8 @@
+## gDRutils 1.1.11 - 2024-04-15
+* add `get_testdata_combo` and `get_testdata_codilution`
+
 ## gDRutils 1.1.10 - 2024-03-07
-- clean up the package
+* clean up the package
 
 ## gDRutils 1.1.9 - 2024-03-07
 * simplify keywords